+++ data/beaconrestapi/build.gradle
@@ -1,6 +1,7 @@
 dependencies {
     implementation project(':data')
     implementation project(':data:provider')
+    implementation project(':data:serializer')
     implementation project(':ethereum:datastructures')
     implementation project(':ethereum:statetransition')
     implementation project(':infrastructure:async')
