+++ services/remote-validator/build.gradle
@@ -1,6 +1,7 @@
 dependencies {
   implementation project(':data:metrics')
   implementation project(':data:provider')
+  implementation project(':data:serializer')
   implementation project(':ethereum:statetransition')
   implementation project(':ethereum:datastructures')
   implementation project(':events')
